Former-commit-id: 8a20ce034265bf595dbfc6593b5ecc13d48b6810

+++ java/src/xmipp/viewer/models/MetadataTableModel.java
@@ -193,7 +193,7 @@ public class MetadataTableModel extends MetadataGalleryTableModel {
 				md.setValueDouble(label, ((Double) value).doubleValue(), id);
 				break;
 			case MetaData.LABEL_SIZET:
-				md.setValueLong(label, ((Long) value).intValue(), id);
+				md.setValueLong(label, ((Long) value).longValue(), id);
 				break;
 			case MetaData.LABEL_STRING:
 			case MetaData.LABEL_VECTOR_DOUBLE:
